+++ include/astra_camera/constants.h
@@ -12,12 +12,12 @@
 
 #pragma once
 
-#include <string>
 #include <cstdlib>
+#include <string>
 
 #define OB_ROS_MAJOR_VERSION 1
 #define OB_ROS_MINOR_VERSION 2
-#define OB_ROS_PATCH_VERSION 5
+#define OB_ROS_PATCH_VERSION 7
 
 #ifndef STRINGIFY
 #define STRINGIFY(arg) #arg
@@ -107,7 +107,11 @@ constexpr static uint32_t GEMIN_E_UVC_PID = 0x055c;
 constexpr static uint32_t GEMINI_E_DEPTH_PID = 0x065c;
 constexpr static uint32_t GEMINI_E_LITE_DEPTH_PID = 0x065d;
 constexpr static uint32_t DABAI_MAX_PID = 0x069a;
+constexpr static uint32_t DABAI_MAX_PRO_PID = 0x069e;
+constexpr static uint32_t GEMINI_UW_PID = 0x06aa;
 constexpr static uint32_t ASTRA_PRO_DEPTH_PID = 0x0403;
+constexpr static uint32_t DABAI_DCW2_DEPTH_PID = 0x06a0;
+constexpr static uint32_t DABAI_DW2_DEPTH_PID = 0x069f;
 const std::string OB_STEREO_S = "Orbbec Canglong";
 const std::string OB_EMBEDDED_S = "Astra SL1000S_U3";
 const std::string OB_STEREO_S_U3 = "Astra SV1301S_U3";