+++ src/ob_camera_node.cpp
@@ -175,8 +175,12 @@ void OBCameraNode::setupDevices() {
   for (const auto& stream_index : IMAGE_STREAMS) {
     if (enable_[stream_index] && !device_->hasSensor(stream_index.first)) {
       std::stringstream ss;
-      ss << "OBCameraNode::setupDevices failed to get sensor:" << stream_name_[stream_index]
-         << " from device. maybe hardware has some problem, or the device is plugged out.";
+      if(stream_index == COLOR && use_uvc_camera_){
+        //Do nothing
+      } else {
+        ss << "OBCameraNode:: Failed to get sensor:" << stream_name_[stream_index]
+           << " from device. maybe hardware has some problem, or the device is not supported.";
+      }
       ROS_WARN_STREAM("check device status " << device_->isValid());
       ROS_WARN_STREAM(ss.str());
     }
@@ -198,8 +202,12 @@ void OBCameraNode::setupDevices() {
       }
       streams_[stream_index] = stream;
     } else {
-      ROS_WARN_STREAM("OBCameraNode::setupDevices failed to create stream "
+      if(stream_index == COLOR && use_uvc_camera_){
+        //Do nothing
+      } else {
+        ROS_WARN_STREAM("OBCameraNode::setupDevices failed to create stream "
                       << stream_name_[stream_index] << ", stream is disabled or sensor not found");
+      }
       if (streams_[stream_index]) {
         ROS_ERROR_STREAM("code should not reach here, it's MUST a hidden bug");
         ROS_WARN_STREAM("OBCameraNode::setupDevices stream " << stream_name_[stream_index]
@@ -209,9 +217,6 @@ void OBCameraNode::setupDevices() {
       enable_[stream_index] = false;
     }
   }
-  device_->setProperty(XN_STREAM_PROPERTY_SOFTWARE_FILTER, soft_filter_);
-  device_->setProperty(XN_STREAM_PROPERTY_DEPTH_MAX_DIFF, soft_filter_max_diff_);
-  device_->setProperty(XN_STREAM_PROPERTY_DEPTH_MAX_SPECKLE_SIZE, soft_filter_max_speckle_size_);
 }
 
 void OBCameraNode::setupFrameCallback() {
@@ -224,6 +229,28 @@ void OBCameraNode::setupFrameCallback() {
       stream_frame_callback_[stream_index] = frame_callback;
     }
   }
+}
+
+void OBCameraNode::setupSyncMode() {
+  OBDEVICE_MODE mode = OBDEVICE_MODE::OBDEVICE_MODE_ROUTINE;
+  std::string mode_str = "none";
+  if (multi_device_sync_mode_ == MultiDeviceSyncMode::Master) {
+    mode = OBDEVICE_MODE::OBDEVICE_MODE_MASTER;
+    mode_str = "master";
+  } else if (multi_device_sync_mode_ == MultiDeviceSyncMode::Slave) {
+    mode = OBDEVICE_MODE::OBDEVICE_MODE_SLAVE;
+    mode_str = "slave";
+  }
+  if (multi_device_sync_mode_ != MultiDeviceSyncMode::None) {
+    ROS_ERROR_STREAM("Set device" << getSerialNumber() << " sync mode to " << mode_str);
+    int value = mode;
+    auto status = device_->setProperty(XN_MODULE_PROPERTY_SET_DEVICE_MODE, value);
+    if (status != openni::STATUS_OK) {
+      ROS_ERROR_STREAM("Set multi device sync mode failed with error "
+                       << openni::OpenNI::getExtendedError());
+    }
+  }
+
 }
 
 void OBCameraNode::setupVideoMode() {
@@ -362,6 +389,17 @@ void OBCameraNode::startStream(const stream_index_pair& stream_index) {
     setIRExposure(ir_exposure_);
   }
 
+  if (stream_index == DEPTH) {
+    streams_[stream_index]->setProperty(XN_STREAM_PROPERTY_SOFTWARE_FILTER, soft_filter_);
+    if (soft_filter_max_diff_ != 0) {
+      streams_[stream_index]->setProperty(XN_STREAM_PROPERTY_DEPTH_MAX_DIFF, soft_filter_max_diff_);
+    }
+    if (soft_filter_max_speckle_size_ != 0) {
+      streams_[stream_index]->setProperty(XN_STREAM_PROPERTY_DEPTH_MAX_SPECKLE_SIZE,
+                                          soft_filter_max_speckle_size_);
+    }
+  }
+
   auto status = streams_[stream_index]->start();
   if (status == openni::STATUS_OK) {
     stream_started_[stream_index] = true;
@@ -458,6 +496,8 @@ void OBCameraNode::getParameters() {
   enable_pointcloud_ = nh_private_.param<bool>("enable_point_cloud", false);
   enable_pointcloud_xyzrgb_ = nh_private_.param<bool>("enable_point_cloud_xyzrgb", false);
   enable_publish_extrinsic_ = nh_private_.param<bool>("enable_publish_extrinsic", false);
+  std::string sync_mode = nh_private_.param<std::string>("multi_device_sync_mode", "none");
+  multi_device_sync_mode_ = getMultiDeviceSyncMode(sync_mode);
   if (depth_align_ && !device_->hasSensor(openni::SENSOR_COLOR) && !use_uvc_camera_) {
     ROS_WARN("No color sensor found, depth align will be disabled");
     depth_align_ = false;
@@ -466,8 +506,8 @@ void OBCameraNode::getParameters() {
     depth_align_ = true;
   }
   soft_filter_ = nh_private_.param<int>("soft_filter", 2);
-  soft_filter_max_diff_ = nh_private_.param<int>("soft_filter_max_diff", 16);
-  soft_filter_max_speckle_size_ = nh_private_.param<int>("soft_filter_max_speckle_size", 480);
+  soft_filter_max_diff_ = nh_private_.param<int>("soft_filter_max_diff", 0);
+  soft_filter_max_speckle_size_ = nh_private_.param<int>("soft_filter_max_speckle_size", 0);
 }
 
 void OBCameraNode::setupTopics() {
@@ -477,6 +517,7 @@ void OBCameraNode::setupTopics() {
   setupDevices();
   setupCameraCtrlServices();
   setupPublishers();
+  setupSyncMode();
   setupVideoMode();
   getCameraParams();
   publishStaticTransforms();
@@ -656,31 +697,49 @@ void OBCameraNode::onNewFrameCallback(const openni::VideoFrameRef& frame,
                                       const stream_index_pair& stream_index) {
   int width = frame.getWidth();
   int height = frame.getHeight();
+  int dst_width = width;
+  int dst_height = height;
   CHECK(images_.count(stream_index));
   auto& image = images_.at(stream_index);
   if (image.size() != cv::Size(width, height)) {
     image.create(height, width, image.type());
   }
   image.data = (uint8_t*)frame.getData();
-  cv::Mat scaled_image;
-  if (stream_index == DEPTH && depth_scale_ != 1) {
-    cv::resize(image, scaled_image, cv::Size(width * depth_scale_, height * depth_scale_), 0, 0,
-               cv::INTER_NEAREST);
+  auto pid = device_info_.getUsbProductId();
+  if (stream_index == DEPTH && depth_align_ && depth_scale_ != 1) {
+    dst_width = width * depth_scale_;
+    dst_height = height * depth_scale_;
+    cv::Mat dst(dst_height, dst_width, CV_16UC1);
+    cv::resize(image, dst, cv::Size(dst_width, dst_height), 0, 0, cv::INTER_NEAREST);
+    image.create(dst_height, dst_width, image.type());
+    image = dst;
+
+  } else if (stream_index == DEPTH && depth_align_ && pid == DABAI_DCW2_DEPTH_PID) {
+    dst_width = width_[COLOR];
+    dst_height = height_[COLOR];
+    cv::Mat dst(dst_height, dst_width, CV_16UC1);
+    dcw2Align(image, dst);
+    image.create(dst_height, dst_width, image.type());
+    image = dst;
+  } else if (stream_index == DEPTH && depth_align_ && (pid == DABAI_MAX_PRO_PID || pid == GEMINI_UW_PID)) {
+    dst_width = width_[COLOR];
+    dst_height = height_[COLOR];
+    cv::Mat dst(dst_height, dst_width, CV_16UC1);
+    maxProAlign(image, dst);
+    image.create(dst_height, dst_width, image.type());
+    image = dst;
   }
   if (flip_image_[stream_index]) {
-    cv::flip(stream_index == DEPTH ? scaled_image : image,
-             stream_index == DEPTH ? scaled_image : image, 1);
+    cv::flip(image, image, 1);
   }
   auto image_msg =
-      *(cv_bridge::CvImage(std_msgs::Header(), encoding_.at(stream_index),
-                           (stream_index == DEPTH && depth_scale_ != 1) ? scaled_image : image)
-            .toImageMsg());
+      *(cv_bridge::CvImage(std_msgs::Header(), encoding_.at(stream_index), image).toImageMsg());
   auto timestamp = ros::Time::now();
   image_msg.header.stamp = timestamp;
   image_msg.header.frame_id =
       depth_align_ ? depth_aligned_frame_id_[stream_index] : optical_frame_id_[stream_index];
-  image_msg.width = (stream_index == DEPTH && depth_scale_ != 1) ? width * depth_scale_ : width;
-  image_msg.height = (stream_index == DEPTH && depth_scale_ != 1) ? height * depth_scale_ : height;
+  image_msg.width = dst_width;
+  image_msg.height = dst_height;
   image_msg.step = image_msg.width * unit_step_size_[stream_index];
   image_msg.is_bigendian = false;
   auto& image_publisher = image_publishers_.at(stream_index);
@@ -707,16 +766,26 @@ void OBCameraNode::onNewFrameCallback(const openni::VideoFrameRef& frame,
     std::string filename = current_path + "/image/" + stream_name_[stream_index] + "_" +
                            std::to_string(image_msg.width) + "x" +
                            std::to_string(image_msg.height) + "_" + std::to_string(fps) + "hz_" +
-                           ss.str() + ".jpg";
+                           ss.str() + ".png";
     if (!boost::filesystem::exists(current_path + "/image")) {
       boost::filesystem::create_directory(current_path + "/image");
     }
     ROS_INFO_STREAM("Saving image to " << filename);
-    if (stream_index != DEPTH) {
+
+    ROS_INFO_STREAM("Saving image to :" << stream_index.first);
+    if (stream_index.first == openni::SENSOR_DEPTH) {
+      auto image_to_save = cv_bridge::toCvCopy(image_msg, encoding_[stream_index])->image;
+      cv::imwrite(filename, image_to_save);
+    } else if (stream_index.first == openni::SENSOR_COLOR) {
+      auto image_to_save =
+          cv_bridge::toCvCopy(image_msg, sensor_msgs::image_encodings::BGR8)->image;
+      cv::imwrite(filename, image_to_save);
+    } else if (stream_index.first == openni::SENSOR_IR) {
       cv::imwrite(filename, image);
     } else {
-      cv::imwrite(filename, scaled_image);
+      ROS_ERROR_STREAM("Unsupported stream type: " << stream_index.first);
     }
+    
     save_images_[stream_index] = false;
   }
 }
@@ -731,7 +800,7 @@ void OBCameraNode::setDepthColorSync(bool data) {
 
 void OBCameraNode::setDepthToColorResolution(int width, int height) {
   const auto pid = device_info_.getUsbProductId();
-  if (pid != DABAI_DCW_DEPTH_PID && pid != GEMINI_E_DEPTH_PID && pid != DABAI_MAX_PID) {
+  if (pid != DABAI_DCW_DEPTH_PID && pid != GEMINI_E_DEPTH_PID) {
     return;
   }
   if (!depth_align_ && !enable_pointcloud_xyzrgb_) {
@@ -891,6 +960,76 @@ void OBCameraNode::pollFrame() {
       continue;
     }
     onNewFrameCallback(frame, stream_index);
+  }
+}
+
+void OBCameraNode::dcw2Align(const cv::Mat& src, cv::Mat& dst) {
+  int color_width = width_[COLOR];
+  int color_height = height_[COLOR];
+  int depth_width = width_[DEPTH];
+  int depth_height = height_[DEPTH];
+  if (color_width == 640 && color_height == 360 && depth_width == 640 && depth_height == 400) {
+    // take src up 360 line
+    cv::Mat src_up = src(cv::Rect(0, 0, 640, 360));
+    dst = src_up;
+  } else if (color_width == 640 && color_height == 360 && depth_width == 540 && depth_height == 400) {
+    // 创建一个新的 640x400 的图像
+    cv::Mat newImage(400, 640, src.type(), cv::Scalar(0));
+    // 定义原始图像在新图像中的位置
+    cv::Rect roi(cv::Rect(50, 0, src.cols, src.rows));
+    // 复制原始图像到新图像的指定位置
+    src.copyTo(newImage(roi));
+
+    cv::Mat src_up = newImage(cv::Rect(0, 0, 640, 360));
+    dst = src_up;
+  } else if (color_width == 640 && color_height == 480 && depth_width == 640 && depth_height == 400) {
+    // take src up 480x360, then upscale to 640x680
+    cv::Mat src_up = src(cv::Rect(80, 0, 480, 360));
+    cv::resize(src_up, dst, cv::Size(640, 480), 0, 0, cv::INTER_NEAREST);
+
+  } else if (color_width == 320 && color_height == 180 && depth_width == 320 &&
+             depth_height == 200) {
+    // take src up 180 line
+    cv::Mat src_up = src(cv::Rect(0, 0, 320, 180));
+    dst = src_up;
+  } else if (color_width == 320 && color_height == 240 && depth_width == 320 &&
+             depth_height == 200) {
+    // take src up 240x180, then upscale to 320x240
+    cv::Mat src_up = src(cv::Rect(40, 0, 240, 180));
+    cv::resize(src_up, dst, cv::Size(320, 240), 0, 0, cv::INTER_NEAREST);
+  } else if (color_width == 640 && color_height == 480 && depth_width == 540 &&
+             depth_height == 400) {
+    // take src up 480x360, then upscale to 640x680
+    cv::Mat src_up = src(cv::Rect(30, 0, 480, 360));
+    cv::resize(src_up, dst, cv::Size(640, 480), 0, 0, cv::INTER_NEAREST);
+  } else if (color_width == 320 && color_height == 240 && depth_width == 270 &&
+             depth_height == 200) {
+    // take src up 240x180, then upscale to 320x240
+    cv::Mat src_up = src(cv::Rect(15, 0, 240, 180));
+    cv::resize(src_up, dst, cv::Size(320, 240), 0, 0, cv::INTER_NEAREST);
+  } else {
+    ROS_ERROR_STREAM("Not support D2C config  color_width "
+                     << color_width << " color_height " << color_height << " depth_width "
+                     << depth_width << " depth_height " << depth_height);
+    return;
+  }
+}
+
+void OBCameraNode::maxProAlign(const cv::Mat& src, cv::Mat& dst) {
+  int color_width = width_[COLOR];
+  int color_height = height_[COLOR];
+  int depth_width = width_[DEPTH];
+  int depth_height = height_[DEPTH];
+  if (color_width == 640 && color_height == 480 && depth_width == 640 && depth_height == 400) {
+    // take src up 480x360, then upscale to 640x680
+    cv::Mat src_up = src(cv::Rect(96, 32, 448, 336));
+    cv::resize(src_up, dst, cv::Size(640, 480), 0, 0, cv::INTER_NEAREST);
+
+  } else {
+    ROS_ERROR_STREAM("Not support D2C config  color_width "
+                     << color_width << " color_height " << color_height << " depth_width "
+                     << depth_width << " depth_height " << depth_height);
+    return;
   }
 }